move validation rule to advanced components
