+++ packages/client/src/pages/Main/containers/EntityDetailBox/EntityDetail/EntityDetailValidationSection/EntityDetailValidationSection.tsx
@@ -19,7 +19,7 @@ import {
   StyledDetailSectionHeader,
   StyledValidationList,
 } from "../EntityDetailStyles";
-import { EntityDetailValidationRule } from "./EntityDetailValidationRule/EntityDetailValidationRule";
+import { ValidationRule } from "components/advanced";
 
 const initValidation: ITerritoryValidation = {
   detail: "",
@@ -171,7 +171,7 @@ export const EntityDetailValidationSection: React.FC<
           {(validations as ITerritoryValidation[]).map((validation, key) => {
             return (
               <React.Fragment key={key}>
-                <EntityDetailValidationRule
+                <ValidationRule
                   key={key}
                   validation={validation}
                   entities={entities}
